+++ src/renderer/components/states/index.ts
@@ -29,7 +29,7 @@ export { HoverContent } from './hover/HoverContent';
 export { ExpandedContent } from './expand/ExpandedContent';
 export { MaxExpandContent } from './maxExpand/MaxExpandContent';
 export { GuideContent } from './guide/GuideContent';
-export { LoginContent } from './login/LoginContent';
+export { LoginContent } from './login';
 export { RegisterContent } from './register/RegisterContent';
 export { PaymentContent } from './payment/PaymentContent';
 export { AnnouncementContent } from './announcement/AnnouncementContent';
