+++ src/renderer/components/states/login/components/LoginForm.tsx
@@ -1,0 +1,146 @@
+import type { ReactElement } from 'react';
+import type { useLogin } from '../hooks/useLogin';
+import { renderFeedback } from '../utils/renderFeedback';
+
+type LoginFormProps = ReturnType<typeof useLogin>;
+
+export function LoginForm(props: LoginFormProps): ReactElement {
+  const {
+    account,
+    setAccount,
+    maskedVerificationEmail,
+    emailCode,
+    setEmailCode,
+    password,
+    setPassword,
+    passwordVisible,
+    setPasswordVisible,
+    submitting,
+    sendingCode,
+    sendCooldownSeconds,
+    needsEmailVerification,
+    isEmailAccount,
+    feedback,
+    handleSendCode,
+    handleSubmit,
+    setRegister,
+    returnFromAuth,
+    t,
+  } = props;
+
+  return (
+    <div className="auth-state-content" onClick={(e) => e.stopPropagation()}>
+      <div className="auth-panel">
+        <div className="auth-panel-title">{t('settings.user.auth.login', { defaultValue: '登录' })}</div>
+        <div className="auth-panel-subtitle">
+          {t('settings.user.auth.hint', { defaultValue: '登录注册服务由 eIsland server 提供' })}
+        </div>
+
+        <label className="settings-field">
+          <span className="settings-field-label">{t('settings.user.fields.account', { defaultValue: '用户名或邮箱' })}</span>
+          <input
+            className="settings-field-input"
+            value={account}
+            onChange={(e) => setAccount(e.target.value)}
+            placeholder={t('settings.user.fields.accountPlaceholder', { defaultValue: '请输入用户名或邮箱' })}
+          />
+        </label>
+
+        <label className="settings-field">
+          <span className="settings-field-label">{t('settings.user.fields.password', { defaultValue: '密码' })}</span>
+          <div className="auth-password-input-wrap">
+            <input
+              className="settings-field-input"
+              type={passwordVisible ? 'text' : 'password'}
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+              placeholder={t('settings.user.fields.passwordPlaceholder', { defaultValue: '至少 8 位，含字母与数字' })}
+            />
+            <button
+              type="button"
+              className="auth-password-toggle"
+              onClick={() => setPasswordVisible((v) => !v)}
+              aria-label={passwordVisible
+                ? t('settings.user.actions.hidePassword', { defaultValue: '隐藏密码' })
+                : t('settings.user.actions.showPassword', { defaultValue: '显示密码' })}
+            >
+              {passwordVisible
+                ? t('settings.user.actions.hide', { defaultValue: '隐藏' })
+                : t('settings.user.actions.show', { defaultValue: '显示' })}
+            </button>
+          </div>
+        </label>
+
+        {needsEmailVerification && (
+          <>
+            {!isEmailAccount && (
+              <label className="settings-field">
+                <span className="settings-field-label">{t('settings.user.fields.boundEmail', { defaultValue: '绑定邮箱' })}</span>
+                <input
+                  className="settings-field-input"
+                  value={maskedVerificationEmail}
+                  readOnly
+                  disabled
+                />
+              </label>
+            )}
+            <label className="settings-field">
+              <span className="settings-field-label">{t('settings.user.fields.emailCode', { defaultValue: '邮箱验证码' })}</span>
+              <div className="auth-password-input-wrap">
+                <input
+                  className="settings-field-input"
+                  value={emailCode}
+                  onChange={(e) => setEmailCode(e.target.value)}
+                  placeholder={t('settings.user.fields.emailCodePlaceholder', { defaultValue: '请输入邮箱验证码' })}
+                />
+                <button
+                  type="button"
+                  className="auth-password-toggle"
+                  onClick={() => void handleSendCode()}
+                  disabled={sendingCode || sendCooldownSeconds > 0}
+                >
+                  {sendingCode
+                    ? t('settings.user.feedback.emailCodeSending', { defaultValue: '发送中…' })
+                    : sendCooldownSeconds > 0
+                      ? t('settings.user.actions.sendCodeCooldown', { defaultValue: '{{seconds}}s后重试', seconds: sendCooldownSeconds })
+                      : t('settings.user.actions.sendCode', { defaultValue: '发送验证码' })}
+                </button>
+              </div>
+            </label>
+          </>
+        )}
+
+        {renderFeedback(feedback)}
+
+        <div className="auth-panel-actions">
+          <button
+            type="button"
+            className="settings-user-primary-btn"
+            onClick={() => void handleSubmit()}
+            disabled={submitting}
+          >
+            {submitting
+              ? t('settings.user.feedback.submitting', { defaultValue: '处理中…' })
+              : t('settings.user.auth.loginBtn', { defaultValue: '登录' })}
+          </button>
+          <button
+            type="button"
+            className="settings-user-secondary-btn"
+            onClick={() => setRegister()}
+            disabled={submitting}
+          >
+            {t('settings.user.auth.register', { defaultValue: '注册' })}
+          </button>
+          <button
+            type="button"
+            className="settings-user-secondary-btn"
+            onClick={returnFromAuth}
+            disabled={submitting}
+          >
+            {t('settings.user.actions.backToCenter', { defaultValue: '返回用户中心' })}
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+}
