+++ src/renderer/components/states/login/index.ts
@@ -1,0 +1,1 @@
+export { LoginContent } from './LoginContent';
