+++ src/renderer/components/states/login/utils/readStandaloneWindowMode.ts
@@ -1,0 +1,18 @@
+import { STANDALONE_WINDOW_MODE_STORE_KEY, LEGACY_COUNTDOWN_WINDOW_MODE_STORE_KEY } from '../config/loginConfig';
+
+export async function readStandaloneWindowMode(): Promise<'integrated' | 'standalone'> {
+  try {
+    const mode = await window.api.storeRead(STANDALONE_WINDOW_MODE_STORE_KEY);
+    if (mode === 'standalone') return 'standalone';
+    if (mode === 'integrated') return 'integrated';
+  } catch {
+    // ignore
+  }
+  try {
+    const legacyMode = await window.api.storeRead(LEGACY_COUNTDOWN_WINDOW_MODE_STORE_KEY);
+    if (legacyMode === 'standalone') return 'standalone';
+  } catch {
+    // ignore
+  }
+  return 'integrated';
+}
